Consolidate migrations
