+++ LICENSE.md
@@ -1,0 +1,5 @@
+# License
+
+This plugin is an official extension of the October CMS platform and is free to use if you have a platform license.
+
+See End User License Agreement at https://octobercms.com/eula for more details.
