+++ updates/version.yaml
@@ -1,8 +1,8 @@
-v1.0.1: Initialize plugin.
+v1.0.1: First version of Location
 v1.0.2:
-    - Create database tables.
-    - create_states_table.php
-    - create_countries_table.php
+    - Create database tables
+    - 000001_create_countries.php
+    - 000002_create_states.php
 v1.0.3:
     - Add seed data for countries and states.
     - seed_all_tables.php
@@ -16,10 +16,9 @@ v1.0.3:
     - seed_pl_states.php
     - seed_dk_states.php
     - seed_ua_states.php
+    - seed_callingcodes.php
 v1.0.4: Satisfy the new Google API key requirement.
-v1.0.5:
-    - Countries can now be pinned to make them appear at the top of the list.
-    - add_country_pinned_flag.php
+v1.0.5: Countries can now be pinned to make them appear at the top of the list.
 v1.0.6: Added support for defining a default country and state.
 v1.0.7: Added basic geocoding method to the Country model.
 v1.0.8: Include Mexico states
@@ -29,14 +28,9 @@ v1.1.2: Yet another change to the AddressFinder for changes to the FormField API
 v1.1.3: Include Argentina states
 v1.1.4: Added support for UK counties
 v1.1.5: Include Italian states (province)
-v1.1.6:
-    - Added ability to disable specific states
-    - add_enabled_states.php
+v1.1.6: Added ability to disable specific states
 v1.2.0: Fixes checkboxes in settings pages
-v1.2.1:
-    - Add country calling codes
-    - add_country_callingcode.php
-    - seed_callingcodes.php
+v1.2.1: Add country calling codes
 v1.2.2: Include Russian states
 v1.2.3: Improve support with October v3
 v1.2.4: Adds reverseStreetNumber option to address finder
